% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anhang.R
\name{get_anhaenge_info}
\alias{get_anhaenge_info}
\alias{download_anhaenge}
\title{Anhänge von eSTATISTIK Nachrichten abrufen und herunterladen}
\usage{
get_anhaenge_info(
  nachrichten_df,
  user_auth,
  verbosity = 0,
  basis_url = "https://erhebungsportal.estatistik.de/Erhebungsportal"
)

download_anhaenge(
  nachrichten_df,
  path = "out/",
  user_auth,
  verbosity = 0,
  basis_url = "https://erhebungsportal.estatistik.de/Erhebungsportal"
)
}
\arguments{
\item{nachrichten_df}{DataFrame der Nachrichten von \code{\link[=get_nachrichten]{get_nachrichten()}}.}

\item{user_auth}{Eine Liste mit den eSTATISTIK Benutzerdaten list(username, password). Das
Passwort wird nachgefragt, wenn es leergelassen wird. Dadurch kann das Speichern des Passworts
im R Skript umgangen werden.}

\item{verbosity}{Default: 0. Detaillevel der Debug Informationen (0 bis 3).}

\item{basis_url}{Die URL des eSATISTIK Erhebungsportals.
Default: https://erhebungsportal.estatistik.de/Erhebungsportal}

\item{path}{Pfad des Ordners indem die heruntergeladenen Anhänge gespeichert werden sollen.}
}
\value{
\code{get_anhaenge_info} gibt ein DataFrame der Nachrichten mit den IDs der Anhänge zurück.
}
\description{
Anhänge von eSTATISTIK Nachrichten abrufen und herunterladen
}
\examples{
\dontrun{
user_auth <- list(username = "IhrBenutzername")
# oder
user_auth <- list(username = "IhrBenutzername", password = "IhrPasswort")

df_nachrichten <- get_nachrichten(user_auth  = user_auth) |>
  get_anhaenge_info(user_auth  = user_auth) |>
  download_anhaenge(user_auth  = user_auth, path = ".")
print(df_nachrichten)
}
}
