% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abonnierte_themen.R
\name{get_abonnierte_themen}
\alias{get_abonnierte_themen}
\title{Abonnierte eSTATISITIK Themen abfragen}
\usage{
get_abonnierte_themen(
  user_auth = list(username = "", password = NULL),
  verbosity = 0,
  basis_url = "https://erhebungsportal.estatistik.de/Erhebungsportal"
)
}
\arguments{
\item{user_auth}{Eine Liste mit den eSTATISTIK Benutzerdaten list(username, password). Das
Passwort wird nachgefragt, wenn es leergelassen wird. Dadurch kann das Speichern des Passworts
im R Skript umgangen werden.}

\item{verbosity}{Default: 0. Detaillevel der Debug Informationen (0 bis 3).}

\item{basis_url}{Die URL des eSATISTIK Erhebungsportals.
Default: https://erhebungsportal.estatistik.de/Erhebungsportal}
}
\value{
Gibt die abonnierten Themen zurück.
}
\description{
Diese Funktion ruft die abonnierten Themen ab.
}
\examples{
\dontrun{
user_auth <- list(username = "IhrBenutzername", password = "IhrPasswort")
get_abonnierte_themen(user_auth = user_auth)

user_auth <- list(username = "IhrBenutzername")
get_abonnierte_themen(user_auth = user_auth)
}
}
