% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nachrichten.R
\name{get_nachrichten}
\alias{get_nachrichten}
\alias{get_nachrichten_raw}
\title{Nachrichten im eSTATSITK Postfach abrufen}
\usage{
get_nachrichten(
  user_auth = list(username = NA, password = NA),
  verbosity = 0,
  basis_url = "https://erhebungsportal.estatistik.de/Erhebungsportal"
)

get_nachrichten_raw(
  user_auth = list(username = NA, password = NA),
  verbosity = 0,
  basis_url = "https://erhebungsportal.estatistik.de/Erhebungsportal"
)
}
\arguments{
\item{user_auth}{Eine Liste mit den eSTATISTIK Benutzerdaten list(username, password). Das
Passwort wird nachgefragt, wenn es leergelassen wird. Dadurch kann das Speichern des Passworts
im R Skript umgangen werden.}

\item{verbosity}{Default: 0. Detaillevel der Debug Informationen (0 bis 3).}

\item{basis_url}{Die URL des eSATISTIK Erhebungsportals.
Default: https://erhebungsportal.estatistik.de/Erhebungsportal}
}
\value{
\code{get_nachrichten} gibt ein DataFrame der Nachrichten zurück.
\code{get_nachrichten_raw} gibt das HTTP-Response-Objekt zurück.
}
\description{
Nachrichten im eSTATSITK Postfach abrufen
}
\examples{
\dontrun{
user_auth <- list(username = "IhrBenutzername")
# oder
user_auth <- list(username = "IhrBenutzername", password = "IhrPasswort")

df_nachrichten <- get_nachrichten(user_auth = user_auth)
print(df_nachrichten)
}
}
