% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markiere_nachrichten_gelesen.R
\name{markiere_nachrichten_gelesen}
\alias{markiere_nachrichten_gelesen}
\title{eSTATISTIK Nachrichten als gelesen markieren.}
\usage{
markiere_nachrichten_gelesen(
  nachrichten_df,
  user_auth = user_auth,
  verbosity = 0,
  basis_url = "https://erhebungsportal.estatistik.de/Erhebungsportal"
)
}
\arguments{
\item{nachrichten_df}{DataFrame der Nachrichten von \code{\link[=get_nachrichten]{get_nachrichten()}}.}

\item{user_auth}{Eine Liste mit den eSTATISTIK Benutzerdaten: \code{list(username = "", password = "")}. Das
Passwort wird nachgefragt, wenn es leergelassen wird. Dadurch kann das Speichern des Passworts
im R Skript umgangen werden.}

\item{verbosity}{Default: 0. Detaillevel der Debug Informationen (0 bis 3).}

\item{basis_url}{Die URL des eSATISTIK Erhebungsportals.
Default: \url{https://erhebungsportal.estatistik.de/Erhebungsportal}}
}
\value{
Gibt eine Liste der API responses zurück.
}
\description{
Gelesene Nachrichten werden in der Nachrichtenliste nicht mehr angezeigt.
}
\examples{
\dontrun{
user_auth <- list(username = "IhrBenutzername")
# oder
user_auth <- list(username = "IhrBenutzername", password = "IhrPasswort")

# Die älteste Nachricht als gelesen markieren
df_nachrichten <- get_nachrichten(user_auth = user_auth) |>
  tail(n = 1) |>
  markiere_nachrichten_gelesen(user_auth = user_auth)
}
}
